+++ tests/Http/Controllers/CP/TaxZoneControllerTest.php
@@ -208,13 +208,3 @@ test('can destroy tax zone and delete assosiated rates', function () {
 
     $this->assertFileDoesNotExist($taxRate->path());
 });
-
-// Helpers
-function user()
-{
-    return User::make()
-        ->makeSuper()
-        ->email('joe.bloggs@example.com')
-        ->set('password', 'secret')
-        ->save();
-}
